refactor(web/applications): move Azure project team code to common
Also includes some structural refactoring of the functions for code
clarity.

refactor(web/applications): move getManyProjectTeamMembersInfo to common
