+++ apps/web/src/server/applications/case/project-team/applications-project-team.service.js
@@ -1,6 +1,6 @@
 import pino from '../../../lib/logger.js';
 import { get, patch, post } from '../../../lib/request.js';
-import projectTeamADService from './application-project-team.azure-service.js';
+import { getProjectTeam } from '../../common/services/project-team.service.js';
 
 /** @typedef {import("../../../app/auth/auth-session.service.js").SessionWithAuth} SessionWithAuth */
 /** @typedef {import('../../applications.types').ProjectTeamMember} ProjectTeamMember */
@@ -50,7 +50,7 @@ export const searchProjectTeamMembers = async (searchTerm, allAzureUsers, pageNu
  */
 export const getProjectTeamMembers = async (caseId) => {
 	try {
-		return { projectTeamMembers: await get(`applications/${caseId}/project-team`) };
+		return { projectTeamMembers: await getProjectTeam(caseId) };
 	} catch (/** @type {*} */ error) {
 		pino.error(`[API] ${error?.response?.body?.error?.code || 'Unknown error'}`);
 
@@ -119,34 +119,4 @@ export const updateProjectTeamMemberRole = async (caseId, userId, role) => {
 
 		return { errors: { query: 'The role could not be saved, try again.' } };
 	}
-};
-
-/**
- * Add extra info (name and email) to the internally stored data of team members (id and role)
- *
- * @param {{userId: string, role: string}[]} projectTeamMembers
- * @param {SessionWithAuth} session
- * @returns {Promise<Partial<ProjectTeamMember>[]>}
- */
-export const getManyProjectTeamMembersInfo = async (projectTeamMembers, session) => {
-	// retrieve all the AD users or throw error
-	// this list contains extra info such as names or emails
-	const allAzureUsers = await projectTeamADService.getAllCachedUsers(session);
-
-	// if for some reason no user can be retrieved from Azure, just return an empty array
-	if (allAzureUsers.length === 0) {
-		return [];
-	}
-
-	// merge the info retrieved from Azure to the internally stored data
-	const projectTeamMembersInfo = projectTeamMembers.map((teamMember) => {
-		const teamMemberInfo = allAzureUsers.find((azureUser) => azureUser.id === teamMember.userId);
-
-		return {
-			...teamMember,
-			...teamMemberInfo
-		};
-	});
-
-	return projectTeamMembersInfo;
 };
